+++ mind_your_stonks/web_driver.py
@@ -1,13 +1,29 @@
+import time
 import psutil
-from psutil import NoSuchProcess
 
 from loguru import logger
 from selenium import webdriver
 from selenium.common.exceptions import TimeoutException
 from selenium.webdriver.firefox.options import Options
 
-
 TIMEOUT = 60.00
+
+
+def disable_images_firefox_profile():
+    """Summary
+    Returns:
+        Object: FirefoxProfile
+    """
+    # Get the Firefox profile object
+    firefoxProfile = webdriver.FirefoxProfile()
+    # Disable images
+    firefoxProfile.set_preference("permissions.default.image", 2)
+    # Disable Flash
+    firefoxProfile.set_preference(
+        "dom.ipc.plugins.enabled.libflashplayer.so", "false"
+    )
+    # Set the modified profile while creating the browser object
+    return firefoxProfile
 
 
 class WebDriverSetup(object):
@@ -16,7 +32,7 @@ class WebDriverSetup(object):
         self._timeout = TIMEOUT
         self._options = Options()
         self._options.headless = headless
-        self._profile = self._disable_images_firefox_profile()
+        self._profile = disable_images_firefox_profile()
         self.logger = logger
 
         self.driver = webdriver.Firefox(
@@ -36,36 +52,17 @@ class WebDriverSetup(object):
             self.logger.exception("Timed-out while loading page.")
             self.close_session()
 
-    def _disable_images_firefox_profile(self):
-        """Summary
-        Returns:
-            Object: FirefoxProfile
-        """
-        # get the Firefox profile object
-        firefoxProfile = webdriver.FirefoxProfile()
-        # Disable images
-        firefoxProfile.set_preference("permissions.default.image", 2)
-        # Disable Flash
-        firefoxProfile.set_preference(
-            "dom.ipc.plugins.enabled.libflashplayer.so", "false"
-        )
-        # Set the modified profile while creating the browser object
-        return firefoxProfile
-
     def close_session(self):
         """Close browser and cleanup"""
         self.logger.info("Closing the browser...")
         self.driver.close()
         self.driver.quit()
+        time.sleep(10)
         PROCNAME = "geckodriver"
-        self.logger.info(f"Cleaning up by killing {PROCNAME} process.")
-        try:
-            _ = [
-                 proc.terminate()
-                 for proc in psutil.process_iter()
-                 if proc.name() == PROCNAME
-            ]
-        except NoSuchProcess:
-            self.logger.debug(f"Process named {PROCNAME} process does not exist.")
-
+        self.logger.info("Cleaning up by killing {} process", PROCNAME)
+        _ = [
+            proc.terminate()
+            for proc in psutil.process_iter()
+            if proc.name() == PROCNAME
+        ]
         self.logger.info("Done...")